+++ src/main/java/com/mudassir/eka/application/generation/GenerationException.java
@@ -1,0 +1,14 @@
+package com.mudassir.eka.application.generation;
+
+import com.mudassir.eka.application.shared.ApplicationException;
+
+public class GenerationException extends ApplicationException {
+
+    public GenerationException(String message) {
+        super(message);
+    }
+
+    public GenerationException(String message, Throwable cause) {
+        super(message, cause);
+    }
+}