+++ src/main/java/com/mudassir/eka/domain/generation/model/GuardrailStatus.java
@@ -1,0 +1,6 @@
+package com.mudassir.eka.domain.generation.model;
+
+public enum GuardrailStatus {
+    PASS,
+    BLOCK
+}