+++ src/main/java/com/mudassir/eka/domain/generation/port/LlmPort.java
@@ -1,0 +1,9 @@
+package com.mudassir.eka.domain.generation.port;
+
+import com.mudassir.eka.domain.generation.model.LlmRequest;
+import com.mudassir.eka.domain.generation.model.LlmResponse;
+
+public interface LlmPort {
+
+    LlmResponse generate(LlmRequest request);
+}